Add worker ID tags in test log
This technique relies on the ability to set log tags that are shared
among all threads of a process. Every test worker worker sets a tag
containing its worker ID just after forking.

This solution works when tests are parallelized using processes, but
fails when using threads for test parallelism.